requirements.txt
asgiref==3.11.0
black==25.11.0
click==8.3.1
colorama==0.4.6
Django==6.0
djangorestframework==3.16.1
django-cors-headers>=4.0
drf-spectacular>=0.27
drf-spectacular-sidecar>=2024.0
ruff>=0.6.0

librt==0.7.3
mccabe==0.7.0
mypy==1.19.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.1
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
python-dotenv==1.2.1
pytokens==0.3.0
setuptools==80.9.0
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
wheel==0.45.1

celery